#38-add-second-screen-to-game-part/add main screen, info screen
